fix: DEFAULT_SEARCH_CONCURRENT_COUNT should be i32 to match concurrent_count field type

+++ src/api/handlers/book.rs
@@ -38,7 +38,7 @@ const DEFAULT_AVAILABLE_CONCURRENT_COUNT: usize = 8;
 /// Per-source search timeout to prevent slow/blocked sources from stalling the whole batch.
 const SEARCH_SOURCE_TIMEOUT_SECS: u64 = 10;
 /// Default concurrent search count (lowered from 24 to reduce network pressure).
-const DEFAULT_SEARCH_CONCURRENT_COUNT: usize = 10;
+const DEFAULT_SEARCH_CONCURRENT_COUNT: i32 = 10;
 const MAX_AVAILABLE_CONCURRENT_COUNT: usize = 20;
 const AVAILABLE_SOURCE_SSE_RESULT_LIMIT: usize = 5;
 const DEFAULT_GLOBAL_EXPLORE_LIMIT: usize = 20;
